+++ providers/arangodb/docs/changelog.rst
@@ -28,6 +28,19 @@
 Changelog
 ---------
 
+2.9.6
+.....
+
+Misc
+~~~~
+
+* ``Use ValueError when no ArangoDB collection operation is specified (#70539)``
+
+.. Below changes are excluded from the changelog. Move them to
+   appropriate section above if needed. Do not delete the lines(!):
+   * ``Use common.compat.sdk for timezone imports in providers (#70492)``
+   * ``Fix inconsistency between generated provider docs and pyproject.toml (#68991)``
+
 2.9.5
 .....